third_party/WebKit/LayoutTests/flag-specific/enable-features=NetworkService/external/wpt/FileAPI/url/cross-global-revoke.sub-expected.txt

This is a testharness.js-based test.
PASS It is possible to revoke same-origin blob URLs from different frames.
PASS It is possible to revoke same-origin blob URLs from a different worker global.
PASS It is not possible to revoke cross-origin blob URLs.
Harness: the test ran to completion.
